Tests added for all methods of Car and Driver CRUD repositories

+++ src/main/java/com/example/demo/dto/CarDTO.java
@@ -4,9 +4,9 @@ import lombok.Data;
 
 @Data
 public class CarDTO {
-    private Long id;
+    private Integer id;
     private String modelName;
     private String carName;
     private String description;
-    private Long driverId;
+    private Integer driverId;
 }
